+++ ovp_search/tests/test_views.py
@@ -175,7 +175,7 @@ class ProjectSearchTestCase(TestCase):
     """
     Test searching with name filter returns project filtered by name(ngram)
     """
-    response = self.client.get(reverse("search-projects-list") + "?name=roject2", format="json")
+    response = self.client.get(reverse("search-projects-list") + "?name=project2", format="json")
     self.assertEqual(len(response.data["results"]), 1)
 
   def test_highlighted_filter(self):
@@ -313,7 +313,7 @@ class OrganizationSearchTestCase(TestCase):
     """
     Test searching with name filter returns organizations filtered by name(ngram)
     """
-    response = self.client.get(reverse("search-organizations-list") + "?name=rganization2", format="json")
+    response = self.client.get(reverse("search-organizations-list") + "?name=organization2", format="json")
     self.assertEqual(len(response.data["results"]), 1)
 
   def test_highlighted_filter(self):
@@ -436,7 +436,7 @@ class UserSearchTestCase(TestCase):
     """
     Test searching with name filter returns organizations filtered by name(ngram)
     """
-    response = self.client.get(reverse("search-users-list") + "?name=ser", format="json")
+    response = self.client.get(reverse("search-users-list") + "?name=user", format="json")
     self.assertEqual(len(response.data["results"]), 3)
 
     response = self.client.get(reverse("search-users-list") + "?name=one", format="json")